[refactor] Presentation Layer View 구조 개선 #1
