+++ CasualConversation/Module/PresentationLayer/Sources/Internal/ViewData/Constant/Language.swift
@@ -1,0 +1,12 @@
+//
+//  Language.swift
+//  Presentation
+//
+//  Created by 김용우 on 6/15/24.
+//  Copyright © 2024 pseapplications. All rights reserved.
+//
+
+enum Language: Int, CaseIterable, Hashable {
+    case original = 0
+    case translation
+}
