+++ CasualConversation/Module/PresentationLayer/Sources/Internal/ViewData/Constant/Tab.swift
@@ -1,0 +1,12 @@
+//
+//  Tab.swift
+//  Presentation
+//
+//  Created by 김용우 on 6/15/24.
+//  Copyright © 2024 pseapplications. All rights reserved.
+//
+
+enum Tab {
+    case conversations
+    case notes
+}
